Vector.svg_object isolated + test
