+++ lib/get_options.js
@@ -1,0 +1,35 @@
+var BITMAP_OPTIONS = {
+    parse_url: false,
+    ec_level: 'M',
+    size: 5,
+    margin: 4,
+    customize: null
+};
+
+var VECTOR_OPTIONS = {
+    parse_url: false,
+    ec_level: 'M',
+    margin: 1,
+    size: 0
+};
+
+function get_options(options, force_type) {
+    if (typeof options === 'string') {
+        options = { 'ec_level': options }
+    } else {
+        options = options || {};
+    }
+    var _options = {
+        type: String(force_type || options.type || 'png').toLowerCase()
+    };
+
+    var defaults = _options.type == 'png' ? BITMAP_OPTIONS : VECTOR_OPTIONS;
+
+    for (var k in defaults) {
+        _options[k] = k in options ? options[k] : defaults[k];
+    }
+
+    return _options;
+}
+
+module.exports = get_options
