+++ lib/qr.js
@@ -8,39 +8,7 @@ var vector = require('./vector');
 
 var fn_noop = function() {};
 
-var BITMAP_OPTIONS = {
-    parse_url: false,
-    ec_level: 'M',
-    size: 5,
-    margin: 4,
-    customize: null
-};
-
-var VECTOR_OPTIONS = {
-    parse_url: false,
-    ec_level: 'M',
-    margin: 1,
-    size: 0
-};
-
-function get_options(options, force_type) {
-    if (typeof options === 'string') {
-        options = { 'ec_level': options }
-    } else {
-        options = options || {};
-    }
-    var _options = {
-        type: String(force_type || options.type || 'png').toLowerCase()
-    };
-
-    var defaults = _options.type == 'png' ? BITMAP_OPTIONS : VECTOR_OPTIONS;
-
-    for (var k in defaults) {
-        _options[k] = k in options ? options[k] : defaults[k];
-    }
-
-    return _options;
-}
+var get_options = require('./get_options')
 
 function qr_image(text, options) {
     options = get_options(options);
@@ -106,16 +74,9 @@ function qr_image_sync(text, options) {
     return result;
 }
 
-function svg_object(text, options) {
-    options = get_options(options, 'svg');
-
-    var matrix = QR(text, options.ec_level);
-    return vector.svg_object(matrix, options.margin);
-}
-
 module.exports = {
     matrix: QR,
     image: qr_image,
     imageSync: qr_image_sync,
-    svgObject: svg_object
+    svgObject: vector.svg_object
 };
