+++ lib/vector.js
@@ -1,5 +1,8 @@
 "use strict";
 
+var QR = require('./qr-base').QR;
+var get_options = require('./get_options')
+
 function matrix2path(matrix) {
     var N = matrix.length;
     var filled = [];
@@ -252,9 +255,16 @@ function PDF(matrix, stream, margin) {
     stream.push(null);
 }
 
+function svg_object(text, options) {
+    options = get_options(options, 'svg');
+
+    var matrix = QR(text, options.ec_level);
+    return SVG_object(matrix, options.margin);
+}
+
 module.exports = {
     svg: SVG,
     eps: EPS,
     pdf: PDF,
-    svg_object: SVG_object
+    svg_object: svg_object
 }
